Fix lag on Linux when using Timer Mod
When Timer Mod was installed and you hit a metal object you would get
massive lag spikes and the sound and particle effect for hitting it
after the first time wouldn't play. This build of the modding api fix
it. This was caused because the System.Runtime.Serialization.dll
couldn't construct some time which it used from other System dlls.